mm: slub: add panic_on_error to the debug facilities
Sometimes it is desirable to override SLUB's debug facilities default
behavior upon stumbling on a cache or object error and just stop the
execution in order to grab a coredump, at the error-spotting time, instead
of trying to fix the issue and report in an attempt to keep the system
rolling.

This patch introduces a new debug flag SLAB_PANIC_ON_ERROR, along with its
related SLUB-machinery, in order to extend current slub_debug facilites
and provide the aforementioned behavior override.

Link: http://lkml.kernel.org/r/20200501211540.71216-1-aquini@redhat.com
Signed-off-by: Rafael Aquini <aquini@redhat.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>